refactor(analyzer): extract external-node helper into call-graph-external.ts (slice 3b)
Next safe slice of modularize-call-graph-builder. Moves the EXTERNAL NODE HELPER
section (classifyExternal, the EXTERNAL_* regex/set tables, getOrCreateExternalNode)
out of call-graph.ts into call-graph-external.ts — pure external-call classification
+ leaf-node interning, depending only on the ExternalKind/FunctionNode types.

All file-internal helpers (never on call-graph.ts's public surface): only
getOrCreateExternalNode is imported back; classifyExternal + the tables stay private
to the new module. Dropped the now-unused internal ExternalKind import (it is still
re-exported on the barrel — surface unchanged). call-graph.ts: 4,951 → 4,887 lines.

Behavior-preserving, verified four ways:
- Snapshot oracle first extended to exercise external nodes across externalKind
  http/db/unknown, then captured before/after: IDENTICAL — SHA-256 3a118017…
- Public export surface byte-identical (multi-line-aware diff: 0 added, 0 removed;
  ExternalKind still exported).
- tsc compiles all importers clean; build/lint/typecheck green.
- Full suite green: 279 files / 5534 tests.

No feature, no dependency, no LLM, no persisted artifact.
